OH2-253 | Reorganize user related endpoints (#463)
* refactor(OH2-253): Reorganize user endpoints

* styles: Reformat code

* chore: Reformat code and improve logging

* chore: Improve logging

* test: Add user update tests

* tests: Update user controller tests

* fix: Fix user controller tests

* chore: Remove .gitkeep from git and remove unnecessary scanBasePackages value

* chore: Add licence header to user controller test and reorganize imports

* chore: Remove useless constructor in UserController

* fix: Fix user controller tests not starting

* chore: Update documentation and updateUser endpoint security

* fix: Remove password from dto return from 'GET /users and /users/{username}'

* update: Allow logged in user to get only its one setting using /users/settings/{id}

* chore: Update endpoint authorization and fix tests

* fix: Prevent user to change its group through update profile endpoint

* fix: Fix api spec file

* refactor: Update user permissions endpoint and remove ResponseEntity like return types

* refactor: Move groups related endpoints into GroupController

* refactor: Move user settings related endpoints into SettingController

* refactor: Refactor packages

* fix: Fix imports, javadoc and use group controller endpoints

* fix: Fix user mapper imports

* fix: Fix UserSettingHelper imports

* chore: Reorganize imports and fix javadoc

* refactor: Refactor user and user groups endpoint and update tests

* fix: Fix user group controller tests and remove menu package

* fix: Fix get user by name and update tests

* chore: Align api spec file

* fix: Encrypt user password

* fix: Fix user group update and associated tests

* chore: Aligne API spec file

* update: Return list of permissions DTO from /users/{username}/permissions endpoint

---------
