+++ src/main/java/org/isf/config/SecurityConfig.java
@@ -21,8 +21,6 @@
  */
 package org.isf.config;
 
-import java.util.Arrays;
-
 import org.isf.permissions.manager.PermissionManager;
 import org.isf.security.ApiAuditorAwareImpl;
 import org.isf.security.CustomLogoutHandler;
@@ -55,29 +53,27 @@ import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.filter.CorsFilter;
 
+import java.util.List;
+
 @Configuration
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(securedEnabled = true)
 public class SecurityConfig {
 
-	@Autowired
-	private UserDetailsService userDetailsService;
-
 	private final TokenProvider tokenProvider;
-
-	@Autowired
-	private RestAuthenticationEntryPoint restAuthenticationEntryPoint;
-
 	@Autowired
 	protected PermissionManager permissionManager;
+	@Autowired
+	private UserDetailsService userDetailsService;
+	@Autowired
+	private RestAuthenticationEntryPoint restAuthenticationEntryPoint;
+	@Autowired
+	private CustomLogoutHandler customLogoutHandler;
 
 	public SecurityConfig(TokenProvider tokenProvider, PermissionManager permissionManager) {
 		this.tokenProvider = tokenProvider;
 		this.permissionManager = permissionManager;
 	}
-
-	@Autowired
-	private CustomLogoutHandler customLogoutHandler;
 
 	@Bean
 	public DaoAuthenticationProvider authenticationProvider() {
@@ -100,7 +96,7 @@ public class SecurityConfig {
 		config.addAllowedHeader("*");
 		// config.setAllowedHeaders(Arrays.asList("Accept", "Accept-Encoding", "Accept-Language", "Authorization", "Content-Type", "Cache-Control",
 		// "Connection", "Cookie", "Host", "Pragma", "Referer, User-Agent"));
-		config.setAllowedMethods(Arrays.asList("*"));
+		config.setAllowedMethods(List.of("*"));
 		// config.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE"));
 		config.setAllowCredentials(true);
 		config.addAllowedOriginPattern("*");
@@ -112,231 +108,233 @@ public class SecurityConfig {
 	@Bean
 	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
 		http.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-						.and().cors()
-						.and().csrf().disable().authorizeRequests()
-						// .expressionHandler(webExpressionHandler())
-						.and().exceptionHandling()
-						// .accessDeniedHandler(accessDeniedHandler)
-						.authenticationEntryPoint(restAuthenticationEntryPoint)
-						.and().authorizeRequests().requestMatchers("/auth/**").permitAll()
+			.and().cors()
+			.and().csrf().disable().authorizeRequests()
+			// .expressionHandler(webExpressionHandler())
+			.and().exceptionHandling()
+			// .accessDeniedHandler(accessDeniedHandler)
+			.authenticationEntryPoint(restAuthenticationEntryPoint)
+			.and().authorizeRequests().requestMatchers("/auth/**").permitAll()
 
-						// admissions
-						.requestMatchers(HttpMethod.POST, "/admissions/**").hasAuthority("admissions.create")
-						.requestMatchers(HttpMethod.GET, "/admissions/**").hasAnyAuthority("admissions.read")
-						.requestMatchers(HttpMethod.PUT, "/admissions/**").hasAuthority("admissions.update")
-						.requestMatchers(HttpMethod.DELETE, "/admissions/**").hasAuthority("admissions.delete")
-						// admissiontypes
-						.requestMatchers(HttpMethod.POST, "/admissiontypes/**").hasAuthority("admissiontypes.create")
-						.requestMatchers(HttpMethod.GET, "/admissiontypes/**").hasAnyAuthority("admissiontypes.read")
-						.requestMatchers(HttpMethod.PUT, "/admissiontypes/**").hasAuthority("admissiontypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/admissiontypes/**").hasAuthority("admissiontypes.delete")
-						// age types
-						.requestMatchers(HttpMethod.GET, "/agetypes/**").hasAnyAuthority("agetypes.read")
-						.requestMatchers(HttpMethod.PUT, "/agetypes/**").hasAuthority("agetypes.update")
-						// dischargetypes
-						.requestMatchers(HttpMethod.POST, "/dischargetypes/**").hasAuthority("dischargetypes.create")
-						.requestMatchers(HttpMethod.GET, "/dischargetypes/**").hasAnyAuthority("dischargetypes.read")
-						.requestMatchers(HttpMethod.PUT, "/dischargetypes/**").hasAuthority("dischargetypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/dischargetypes/**").hasAuthority("dischargetypes.delete")
-						// diseases
-						.requestMatchers(HttpMethod.POST, "/diseases/**").hasAuthority("diseases.create")
-						.requestMatchers(HttpMethod.GET, "/diseases/**").hasAnyAuthority("diseases.read")
-						.requestMatchers(HttpMethod.PUT, "/diseases/**").hasAuthority("diseases.update")
-						.requestMatchers(HttpMethod.DELETE, "/diseases/**").hasAuthority("diseases.delete")
-						// diseasetypes
-						.requestMatchers(HttpMethod.POST, "/diseasetypes/**").hasAuthority("diseasetypes.create")
-						.requestMatchers(HttpMethod.GET, "/diseasetypes/**").hasAnyAuthority("diseasetypes.read")
-						.requestMatchers(HttpMethod.PUT, "/diseasetypes/**").hasAuthority("diseasetypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/diseasetypes/**").hasAuthority("diseasetypes.delete")
-						// deliveryresulttype
-						.requestMatchers(HttpMethod.POST, "/deliveryresulttypes/**").hasAuthority("deliveryresulttypes.create")
-						.requestMatchers(HttpMethod.GET, "/deliveryresulttypes/**").hasAnyAuthority("deliveryresulttypes.read")
-						.requestMatchers(HttpMethod.PUT, "/deliveryresulttypes/**").hasAuthority("deliveryresulttypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/deliveryresulttypes/**").hasAuthority("deliveryresulttypes.delete")
-						// deliverytypes
-						.requestMatchers(HttpMethod.POST, "/deliverytypes/**").hasAuthority("deliverytypes.create")
-						.requestMatchers(HttpMethod.GET, "/deliverytypes/**").hasAnyAuthority("deliverytypes.read")
-						.requestMatchers(HttpMethod.PUT, "/deliverytypes/**").hasAuthority("deliverytypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/deliverytypes/**").hasAuthority("deliverytypes.delete")
-						// exams
-						.requestMatchers(HttpMethod.POST, "/exams/**").hasAuthority("exams.create")
-						.requestMatchers(HttpMethod.GET, "/exams/**").hasAnyAuthority("exams.read")
-						.requestMatchers(HttpMethod.PUT, "/exams/**").hasAuthority("exams.update")
-						.requestMatchers(HttpMethod.DELETE, "/exams/**").hasAuthority("exams.delete")
-						// examrows
-						.requestMatchers(HttpMethod.POST, "/examrows/**").hasAuthority("examrows.create")
-						.requestMatchers(HttpMethod.GET, "/examrows/**").hasAnyAuthority("examrows.read")
-						.requestMatchers(HttpMethod.PUT, "/examrows/**").hasAuthority("examrows.update")
-						.requestMatchers(HttpMethod.DELETE, "/examrows/**").hasAuthority("examrows.delete")
-						// examinations
-						.requestMatchers(HttpMethod.POST, "/examinations/**").hasAuthority("examinations.create")
-						.requestMatchers(HttpMethod.GET, "/examinations/**").hasAnyAuthority("examinations.read")
-						.requestMatchers(HttpMethod.PUT, "/examinations/**").hasAuthority("examinations.update")
-						.requestMatchers(HttpMethod.DELETE, "/examinations/**").hasAuthority("examinations.delete")
-						// examtypes
-						.requestMatchers(HttpMethod.POST, "/examtypes/**").hasAuthority("examtypes.create")
-						.requestMatchers(HttpMethod.GET, "/examtypes/**").hasAnyAuthority("examtypes.read")
-						.requestMatchers(HttpMethod.PUT, "/examtypes/**").hasAuthority("examtypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/examtypes/**").hasAuthority("examtypes.delete")
-						// hospitals
-						.requestMatchers(HttpMethod.POST, "/hospitals/**").hasAuthority("hospitals.create")
-						// .requestMatchers(HttpMethod.GET, "/hospitals/**").hasAnyAuthority("hospital.read") to anyone
-						.requestMatchers(HttpMethod.PUT, "/hospitals/**").hasAuthority("hospitals.update")
-						.requestMatchers(HttpMethod.DELETE, "/hospitals/**").hasAuthority("hospitals.delete")
-						// laboratories
-						.requestMatchers(HttpMethod.POST, "/laboratories/**").hasAuthority("laboratories.create")
-						.requestMatchers(HttpMethod.GET, "/laboratories/**").hasAnyAuthority("laboratories.read")
-						.requestMatchers(HttpMethod.PUT, "/laboratories/**").hasAuthority("laboratories.update")
-						.requestMatchers(HttpMethod.DELETE, "/laboratories/**").hasAuthority("laboratories.delete")
-						// malnutrition
-						.requestMatchers(HttpMethod.POST, "/malnutritions/**").hasAuthority("malnutritions.create")
-						.requestMatchers(HttpMethod.GET, "/malnutritions/**").hasAuthority("malnutritions.read")
-						.requestMatchers(HttpMethod.PUT, "/malnutritions/**").hasAuthority("malnutritions.update")
-						.requestMatchers(HttpMethod.DELETE, "/malnutritions/**").hasAuthority("malnutritions.delete")
-						// medicals
-						.requestMatchers(HttpMethod.POST, "/medicals/**").hasAuthority("medicals.create")
-						.requestMatchers(HttpMethod.GET, "/medicals/**").hasAuthority("medicals.read")
-						.requestMatchers(HttpMethod.PUT, "/medicals/**").hasAuthority("medicals.update")
-						.requestMatchers(HttpMethod.DELETE, "/medicals/**").hasAuthority("medicals.delete")
-						// medicalstock
-						.requestMatchers(HttpMethod.POST, "/medicalstockmovements/**").hasAuthority("medicalstockmovements.create")
-						.requestMatchers(HttpMethod.GET, "/medicalstockmovements/**").hasAuthority("medicalstockmovements.read")
-						.requestMatchers(HttpMethod.PUT, "/medicalstockmovements/**").hasAuthority("medicalstockmovements.update")
-						.requestMatchers(HttpMethod.DELETE, "/medicalstockmovements/**").hasAuthority("medicalstockmovements.delete")
-						// medicalstockward
-						.requestMatchers(HttpMethod.POST, "/medicalstockward/**").hasAuthority("medicalstockward.create")
-						.requestMatchers(HttpMethod.GET, "/medicalstockward/**").hasAuthority("medicalstockward.read")
-						.requestMatchers(HttpMethod.PUT, "/medicalstockward/**").hasAuthority("medicalstockward.update")
-						.requestMatchers(HttpMethod.DELETE, "/medicalstockward/**").hasAuthority("medicalstockward.delete")
-						// medicalstockmovtype
-						.requestMatchers(HttpMethod.POST, "/medstockmovementtypes/**").hasAuthority("medstockmovementtypes.create")
-						.requestMatchers(HttpMethod.GET, "/medstockmovementtypes/**").hasAuthority("medstockmovementtypes.read")
-						.requestMatchers(HttpMethod.PUT, "/medstockmovementtypes/**").hasAuthority("medstockmovementtypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/medstockmovementtypes/**").hasAuthority("medstockmovementtypes.delete")
-						// medicaltype
-						.requestMatchers(HttpMethod.POST, "/medicaltypes/**").hasAuthority("medicaltypes.create")
-						.requestMatchers(HttpMethod.GET, "/medicaltypes/**").hasAuthority("medicaltypes.read")
-						.requestMatchers(HttpMethod.PUT, "/medicaltypes/**").hasAuthority("medicaltypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/medicaltypes/**").hasAuthority("medicaltypes.delete")
-						// opd
-						.requestMatchers(HttpMethod.POST, "/opds/**").hasAuthority("opds.create")
-						.requestMatchers(HttpMethod.GET, "/opds/**").hasAnyAuthority("opds.read")
-						.requestMatchers(HttpMethod.PUT, "/opds/**").hasAuthority("opds.update")
-						.requestMatchers(HttpMethod.DELETE, "/opds/**").hasAuthority("opds.delete")
-						// operations
-						.requestMatchers(HttpMethod.POST, "/operations/**").hasAuthority("operations.create")
-						.requestMatchers(HttpMethod.GET, "/operations/**").hasAnyAuthority("operations.read")
-						.requestMatchers(HttpMethod.PUT, "/operations/**").hasAuthority("operations.update")
-						.requestMatchers(HttpMethod.DELETE, "/operations/**").hasAuthority("operations.delete")
-						// operation types
-						.requestMatchers(HttpMethod.POST, "/operationtypes/**").hasAuthority("operationtypes.create")
-						.requestMatchers(HttpMethod.GET, "/operationtypes/**").hasAnyAuthority("operationtypes.read")
-						.requestMatchers(HttpMethod.PUT, "/operationtypes/**").hasAuthority("operationtypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/operationtypes/**").hasAuthority("operationtypes.delete")
-						// patientconsensus
-						.requestMatchers(HttpMethod.POST, "/patientconsensus/**").hasAuthority("patientconsensus.create")
-						.requestMatchers(HttpMethod.GET, "/patientconsensus/**").hasAuthority("patientconsensus.read")
-						.requestMatchers(HttpMethod.PUT, "/patientconsensus/**").hasAuthority("patientconsensus.update")
-						.requestMatchers(HttpMethod.DELETE, "/patientconsensus/**").hasAuthority("patientconsensus.delete")
-						// patients
-						.requestMatchers(HttpMethod.POST, "/patients/**").hasAuthority("patients.create")
-						.requestMatchers(HttpMethod.GET, "/patients/**").hasAuthority("patients.read")
-						.requestMatchers(HttpMethod.PUT, "/patients/**").hasAuthority("patients.update")
-						.requestMatchers(HttpMethod.DELETE, "/patients/**").hasAuthority("patients.delete")
-						// patientvaccines
-						.requestMatchers(HttpMethod.POST, "/patientvaccines/**").hasAuthority("patientvaccines.create")
-						.requestMatchers(HttpMethod.GET, "/patientvaccines/**").hasAnyAuthority("patientvaccines.read")
-						.requestMatchers(HttpMethod.PUT, "/patientvaccines/**").hasAuthority("patientvaccines.update")
-						.requestMatchers(HttpMethod.DELETE, "/patientvaccines/**").hasAuthority("patientvaccines.delete")
-						// permission
-						.requestMatchers(HttpMethod.POST, "/permissions/**").hasAuthority("permissions.create")
-						.requestMatchers(HttpMethod.GET, "/permissions/**").hasAuthority("permissions.read")
-						.requestMatchers(HttpMethod.PUT, "/permissions/**").hasAuthority("permissions.update")
-						.requestMatchers(HttpMethod.DELETE, "/permissions/**").hasAuthority("permissions.delete")
-						// grouppermission
-						.requestMatchers(HttpMethod.POST, "/users/groups").hasAuthority("grouppermission.create")
-						.requestMatchers(HttpMethod.GET, "/users/groups/**").hasAuthority("grouppermission.read")
-						.requestMatchers(HttpMethod.PUT, "/users/groups").hasAuthority("grouppermission.update")
-						.requestMatchers(HttpMethod.DELETE, "/users/groups/**").hasAuthority("grouppermission.delete")
-						// user
-						.requestMatchers(HttpMethod.POST, "/users").hasAuthority("users.create")
-						.requestMatchers(HttpMethod.GET, "/users/**").hasAuthority("users.read")
-						.requestMatchers(HttpMethod.PUT, "/users").hasAuthority("users.update")
-						.requestMatchers(HttpMethod.DELETE, "/users/**").hasAuthority("users.delete")
-						// user setting
-						.requestMatchers(HttpMethod.GET, "/users/settings/**").hasAuthority("usersettings.read")
-						.requestMatchers(HttpMethod.GET, "/users/{userName}/settings/{configName}").hasAuthority("usersettings.read")
-						.requestMatchers(HttpMethod.POST, "/users/settings/**").hasAuthority("usersettings.create")
-						.requestMatchers(HttpMethod.PUT, "/users/settings/**").hasAuthority("usersettings.update")
-						.requestMatchers(HttpMethod.DELETE, "/users/settings/**").hasAuthority("usersettings.delete")
-						// pregnanttreatmenttypes
-						.requestMatchers(HttpMethod.POST, "/pregnanttreatmenttypes/**").hasAuthority("pregnanttreatmenttypes.create")
-						.requestMatchers(HttpMethod.GET, "/pregnanttreatmenttypes/**").hasAnyAuthority("pregnanttreatmenttypes.read")
-						.requestMatchers(HttpMethod.PUT, "/pregnanttreatmenttypes/**").hasAuthority("pregnanttreatmenttypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/pregnanttreatmenttypes/**").hasAuthority("pregnanttreatmenttypes.delete")
-						// pricelists
-						.requestMatchers(HttpMethod.POST, "/pricelists/**").hasAuthority("pricelists.create")
-						.requestMatchers(HttpMethod.GET, "/pricelists/**").hasAnyAuthority("pricelists.read")
-						.requestMatchers(HttpMethod.PUT, "/pricelists/**").hasAuthority("pricelists.update")
-						.requestMatchers(HttpMethod.DELETE, "/pricelists/**").hasAuthority("pricelists.delete")
-						// pricesothers
-						.requestMatchers(HttpMethod.POST, "/pricesothers/**").hasAuthority("pricesothers.create")
-						.requestMatchers(HttpMethod.GET, "/pricesothers/**").hasAnyAuthority("pricesothers.read")
-						.requestMatchers(HttpMethod.PUT, "/pricesothers/**").hasAuthority("pricesothers.update")
-						.requestMatchers(HttpMethod.DELETE, "/pricesothers/**").hasAuthority("pricesothers.delete")
-						// reports
-						.requestMatchers(HttpMethod.POST, "/reports/**").hasAuthority("reports.create")
-						.requestMatchers(HttpMethod.GET, "/reports/**").hasAnyAuthority("reports.read")
-						.requestMatchers(HttpMethod.PUT, "/reports/**").hasAuthority("reports.update")
-						.requestMatchers(HttpMethod.DELETE, "/reports/**").hasAuthority("reports.delete")
-						// sms
-						.requestMatchers(HttpMethod.POST, "/sms/**").hasAuthority("sms.create")
-						.requestMatchers(HttpMethod.GET, "/sms/**").hasAnyAuthority("sms.read")
-						.requestMatchers(HttpMethod.PUT, "/sms/**").hasAuthority("sms.update")
-						.requestMatchers(HttpMethod.DELETE, "/sms/**").hasAuthority("sms.delete")
-						// suppliers
-						.requestMatchers(HttpMethod.POST, "/suppliers/**").hasAuthority("suppliers.create")
-						.requestMatchers(HttpMethod.GET, "/suppliers/**").hasAnyAuthority("suppliers.read")
-						.requestMatchers(HttpMethod.PUT, "/suppliers/**").hasAuthority("suppliers.update")
-						.requestMatchers(HttpMethod.DELETE, "/suppliers/**").hasAuthority("suppliers.delete")
-						// therapy
-						.requestMatchers(HttpMethod.POST, "/therapies/**").hasAuthority("therapies.create")
-						.requestMatchers(HttpMethod.GET, "/therapies/**").hasAnyAuthority("therapies.read")
-						.requestMatchers(HttpMethod.PUT, "/therapies/**").hasAuthority("therapies.update")
-						.requestMatchers(HttpMethod.DELETE, "/therapies/**").hasAuthority("therapies.delete")
-						// vaccines
-						.requestMatchers(HttpMethod.POST, "/vaccines/**").hasAuthority("vaccines.create")
-						.requestMatchers(HttpMethod.GET, "/vaccines/**").hasAnyAuthority("vaccines.read")
-						.requestMatchers(HttpMethod.PUT, "/vaccines/**").hasAuthority("vaccines.update")
-						.requestMatchers(HttpMethod.DELETE, "/vaccines/**").hasAuthority("vaccines.delete")
-						// vaccineType
-						.requestMatchers(HttpMethod.POST, "/vaccinetypes/**").hasAuthority("vaccinetypes.create")
-						.requestMatchers(HttpMethod.GET, "/vaccinetypes/**").hasAnyAuthority("vaccinetypes.read")
-						.requestMatchers(HttpMethod.PUT, "/vaccinetypes/**").hasAuthority("vaccinetypes.update")
-						.requestMatchers(HttpMethod.DELETE, "/vaccinetypes/**").hasAuthority("vaccinetypes.delete")
-						// visit
-						.requestMatchers(HttpMethod.POST, "/visits/**").hasAuthority("visits.create")
-						.requestMatchers(HttpMethod.GET, "/visits/**").hasAnyAuthority("visits.read")
-						.requestMatchers(HttpMethod.PUT, "/visits/**").hasAuthority("visits.update")
-						.requestMatchers(HttpMethod.DELETE, "/visits/**").hasAuthority("visits.delete")
-						// wards
-						.requestMatchers(HttpMethod.POST, "/wards/**").hasAuthority("wards.create")
-						.requestMatchers(HttpMethod.GET, "/wards/**").hasAnyAuthority("wards.read")
-						.requestMatchers(HttpMethod.PUT, "/wards/**").hasAuthority("wards.update")
-						.requestMatchers(HttpMethod.DELETE, "/wards/**").hasAuthority("wards.delete")
+			// admissions
+			.requestMatchers(HttpMethod.POST, "/admissions/**").hasAuthority("admissions.create")
+			.requestMatchers(HttpMethod.GET, "/admissions/**").hasAnyAuthority("admissions.read")
+			.requestMatchers(HttpMethod.PUT, "/admissions/**").hasAuthority("admissions.update")
+			.requestMatchers(HttpMethod.DELETE, "/admissions/**").hasAuthority("admissions.delete")
+			// admissiontypes
+			.requestMatchers(HttpMethod.POST, "/admissiontypes/**").hasAuthority("admissiontypes.create")
+			.requestMatchers(HttpMethod.GET, "/admissiontypes/**").hasAnyAuthority("admissiontypes.read")
+			.requestMatchers(HttpMethod.PUT, "/admissiontypes/**").hasAuthority("admissiontypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/admissiontypes/**").hasAuthority("admissiontypes.delete")
+			// age types
+			.requestMatchers(HttpMethod.GET, "/agetypes/**").hasAnyAuthority("agetypes.read")
+			.requestMatchers(HttpMethod.PUT, "/agetypes/**").hasAuthority("agetypes.update")
+			// dischargetypes
+			.requestMatchers(HttpMethod.POST, "/dischargetypes/**").hasAuthority("dischargetypes.create")
+			.requestMatchers(HttpMethod.GET, "/dischargetypes/**").hasAnyAuthority("dischargetypes.read")
+			.requestMatchers(HttpMethod.PUT, "/dischargetypes/**").hasAuthority("dischargetypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/dischargetypes/**").hasAuthority("dischargetypes.delete")
+			// diseases
+			.requestMatchers(HttpMethod.POST, "/diseases/**").hasAuthority("diseases.create")
+			.requestMatchers(HttpMethod.GET, "/diseases/**").hasAnyAuthority("diseases.read")
+			.requestMatchers(HttpMethod.PUT, "/diseases/**").hasAuthority("diseases.update")
+			.requestMatchers(HttpMethod.DELETE, "/diseases/**").hasAuthority("diseases.delete")
+			// diseasetypes
+			.requestMatchers(HttpMethod.POST, "/diseasetypes/**").hasAuthority("diseasetypes.create")
+			.requestMatchers(HttpMethod.GET, "/diseasetypes/**").hasAnyAuthority("diseasetypes.read")
+			.requestMatchers(HttpMethod.PUT, "/diseasetypes/**").hasAuthority("diseasetypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/diseasetypes/**").hasAuthority("diseasetypes.delete")
+			// deliveryresulttype
+			.requestMatchers(HttpMethod.POST, "/deliveryresulttypes/**").hasAuthority("deliveryresulttypes.create")
+			.requestMatchers(HttpMethod.GET, "/deliveryresulttypes/**").hasAnyAuthority("deliveryresulttypes.read")
+			.requestMatchers(HttpMethod.PUT, "/deliveryresulttypes/**").hasAuthority("deliveryresulttypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/deliveryresulttypes/**").hasAuthority("deliveryresulttypes.delete")
+			// deliverytypes
+			.requestMatchers(HttpMethod.POST, "/deliverytypes/**").hasAuthority("deliverytypes.create")
+			.requestMatchers(HttpMethod.GET, "/deliverytypes/**").hasAnyAuthority("deliverytypes.read")
+			.requestMatchers(HttpMethod.PUT, "/deliverytypes/**").hasAuthority("deliverytypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/deliverytypes/**").hasAuthority("deliverytypes.delete")
+			// exams
+			.requestMatchers(HttpMethod.POST, "/exams/**").hasAuthority("exams.create")
+			.requestMatchers(HttpMethod.GET, "/exams/**").hasAnyAuthority("exams.read")
+			.requestMatchers(HttpMethod.PUT, "/exams/**").hasAuthority("exams.update")
+			.requestMatchers(HttpMethod.DELETE, "/exams/**").hasAuthority("exams.delete")
+			// examrows
+			.requestMatchers(HttpMethod.POST, "/examrows/**").hasAuthority("examrows.create")
+			.requestMatchers(HttpMethod.GET, "/examrows/**").hasAnyAuthority("examrows.read")
+			.requestMatchers(HttpMethod.PUT, "/examrows/**").hasAuthority("examrows.update")
+			.requestMatchers(HttpMethod.DELETE, "/examrows/**").hasAuthority("examrows.delete")
+			// examinations
+			.requestMatchers(HttpMethod.POST, "/examinations/**").hasAuthority("examinations.create")
+			.requestMatchers(HttpMethod.GET, "/examinations/**").hasAnyAuthority("examinations.read")
+			.requestMatchers(HttpMethod.PUT, "/examinations/**").hasAuthority("examinations.update")
+			.requestMatchers(HttpMethod.DELETE, "/examinations/**").hasAuthority("examinations.delete")
+			// examtypes
+			.requestMatchers(HttpMethod.POST, "/examtypes/**").hasAuthority("examtypes.create")
+			.requestMatchers(HttpMethod.GET, "/examtypes/**").hasAnyAuthority("examtypes.read")
+			.requestMatchers(HttpMethod.PUT, "/examtypes/**").hasAuthority("examtypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/examtypes/**").hasAuthority("examtypes.delete")
+			// hospitals
+			.requestMatchers(HttpMethod.POST, "/hospitals/**").hasAuthority("hospitals.create")
+			// .requestMatchers(HttpMethod.GET, "/hospitals/**").hasAnyAuthority("hospital.read") to anyone
+			.requestMatchers(HttpMethod.PUT, "/hospitals/**").hasAuthority("hospitals.update")
+			.requestMatchers(HttpMethod.DELETE, "/hospitals/**").hasAuthority("hospitals.delete")
+			// laboratories
+			.requestMatchers(HttpMethod.POST, "/laboratories/**").hasAuthority("laboratories.create")
+			.requestMatchers(HttpMethod.GET, "/laboratories/**").hasAnyAuthority("laboratories.read")
+			.requestMatchers(HttpMethod.PUT, "/laboratories/**").hasAuthority("laboratories.update")
+			.requestMatchers(HttpMethod.DELETE, "/laboratories/**").hasAuthority("laboratories.delete")
+			// malnutrition
+			.requestMatchers(HttpMethod.POST, "/malnutritions/**").hasAuthority("malnutritions.create")
+			.requestMatchers(HttpMethod.GET, "/malnutritions/**").hasAuthority("malnutritions.read")
+			.requestMatchers(HttpMethod.PUT, "/malnutritions/**").hasAuthority("malnutritions.update")
+			.requestMatchers(HttpMethod.DELETE, "/malnutritions/**").hasAuthority("malnutritions.delete")
+			// medicals
+			.requestMatchers(HttpMethod.POST, "/medicals/**").hasAuthority("medicals.create")
+			.requestMatchers(HttpMethod.GET, "/medicals/**").hasAuthority("medicals.read")
+			.requestMatchers(HttpMethod.PUT, "/medicals/**").hasAuthority("medicals.update")
+			.requestMatchers(HttpMethod.DELETE, "/medicals/**").hasAuthority("medicals.delete")
+			// medicalstock
+			.requestMatchers(HttpMethod.POST, "/medicalstockmovements/**").hasAuthority("medicalstockmovements.create")
+			.requestMatchers(HttpMethod.GET, "/medicalstockmovements/**").hasAuthority("medicalstockmovements.read")
+			.requestMatchers(HttpMethod.PUT, "/medicalstockmovements/**").hasAuthority("medicalstockmovements.update")
+			.requestMatchers(HttpMethod.DELETE, "/medicalstockmovements/**").hasAuthority("medicalstockmovements.delete")
+			// medicalstockward
+			.requestMatchers(HttpMethod.POST, "/medicalstockward/**").hasAuthority("medicalstockward.create")
+			.requestMatchers(HttpMethod.GET, "/medicalstockward/**").hasAuthority("medicalstockward.read")
+			.requestMatchers(HttpMethod.PUT, "/medicalstockward/**").hasAuthority("medicalstockward.update")
+			.requestMatchers(HttpMethod.DELETE, "/medicalstockward/**").hasAuthority("medicalstockward.delete")
+			// medicalstockmovtype
+			.requestMatchers(HttpMethod.POST, "/medstockmovementtypes/**").hasAuthority("medstockmovementtypes.create")
+			.requestMatchers(HttpMethod.GET, "/medstockmovementtypes/**").hasAuthority("medstockmovementtypes.read")
+			.requestMatchers(HttpMethod.PUT, "/medstockmovementtypes/**").hasAuthority("medstockmovementtypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/medstockmovementtypes/**").hasAuthority("medstockmovementtypes.delete")
+			// medicaltype
+			.requestMatchers(HttpMethod.POST, "/medicaltypes/**").hasAuthority("medicaltypes.create")
+			.requestMatchers(HttpMethod.GET, "/medicaltypes/**").hasAuthority("medicaltypes.read")
+			.requestMatchers(HttpMethod.PUT, "/medicaltypes/**").hasAuthority("medicaltypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/medicaltypes/**").hasAuthority("medicaltypes.delete")
+			// opd
+			.requestMatchers(HttpMethod.POST, "/opds/**").hasAuthority("opds.create")
+			.requestMatchers(HttpMethod.GET, "/opds/**").hasAnyAuthority("opds.read")
+			.requestMatchers(HttpMethod.PUT, "/opds/**").hasAuthority("opds.update")
+			.requestMatchers(HttpMethod.DELETE, "/opds/**").hasAuthority("opds.delete")
+			// operations
+			.requestMatchers(HttpMethod.POST, "/operations/**").hasAuthority("operations.create")
+			.requestMatchers(HttpMethod.GET, "/operations/**").hasAnyAuthority("operations.read")
+			.requestMatchers(HttpMethod.PUT, "/operations/**").hasAuthority("operations.update")
+			.requestMatchers(HttpMethod.DELETE, "/operations/**").hasAuthority("operations.delete")
+			// operation types
+			.requestMatchers(HttpMethod.POST, "/operationtypes/**").hasAuthority("operationtypes.create")
+			.requestMatchers(HttpMethod.GET, "/operationtypes/**").hasAnyAuthority("operationtypes.read")
+			.requestMatchers(HttpMethod.PUT, "/operationtypes/**").hasAuthority("operationtypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/operationtypes/**").hasAuthority("operationtypes.delete")
+			// patientconsensus
+			.requestMatchers(HttpMethod.POST, "/patientconsensus/**").hasAuthority("patientconsensus.create")
+			.requestMatchers(HttpMethod.GET, "/patientconsensus/**").hasAuthority("patientconsensus.read")
+			.requestMatchers(HttpMethod.PUT, "/patientconsensus/**").hasAuthority("patientconsensus.update")
+			.requestMatchers(HttpMethod.DELETE, "/patientconsensus/**").hasAuthority("patientconsensus.delete")
+			// patients
+			.requestMatchers(HttpMethod.POST, "/patients/**").hasAuthority("patients.create")
+			.requestMatchers(HttpMethod.GET, "/patients/**").hasAuthority("patients.read")
+			.requestMatchers(HttpMethod.PUT, "/patients/**").hasAuthority("patients.update")
+			.requestMatchers(HttpMethod.DELETE, "/patients/**").hasAuthority("patients.delete")
+			// patientvaccines
+			.requestMatchers(HttpMethod.POST, "/patientvaccines/**").hasAuthority("patientvaccines.create")
+			.requestMatchers(HttpMethod.GET, "/patientvaccines/**").hasAnyAuthority("patientvaccines.read")
+			.requestMatchers(HttpMethod.PUT, "/patientvaccines/**").hasAuthority("patientvaccines.update")
+			.requestMatchers(HttpMethod.DELETE, "/patientvaccines/**").hasAuthority("patientvaccines.delete")
+			// permission
+			.requestMatchers(HttpMethod.POST, "/permissions/**").hasAuthority("permissions.create")
+			.requestMatchers(HttpMethod.GET, "/permissions/**").hasAuthority("permissions.read")
+			.requestMatchers(HttpMethod.PUT, "/permissions/**").hasAuthority("permissions.update")
+			.requestMatchers(HttpMethod.DELETE, "/permissions/**").hasAuthority("permissions.delete")
+			// grouppermission
+			.requestMatchers(HttpMethod.POST, "/usergroups/{group_code}/permissions/**").hasAuthority("grouppermission.create")
+			.requestMatchers(HttpMethod.GET, "/usergroups/{group_code}/permissions/**").hasAuthority("grouppermission.read")
+			.requestMatchers(HttpMethod.PUT, "/usergroups/{group_code}/permissions/**").hasAuthority("grouppermission.update")
+			.requestMatchers(HttpMethod.DELETE, "/usergroups/{group_code}/permissions/**").hasAuthority("grouppermission.delete")
+			// usergroups
+			.requestMatchers(HttpMethod.POST, "/usergroups/**").hasAuthority("usergroups.create")
+			.requestMatchers(HttpMethod.GET, "/usergroups/**").hasAuthority("usergroups.read")
+			.requestMatchers(HttpMethod.PUT, "/usergroups/**").hasAuthority("usergroups.update")
+			.requestMatchers(HttpMethod.DELETE, "/usergroups/**").hasAuthority("usergroups.delete")
+			// user
+			.requestMatchers("/users/me").authenticated()
+			.requestMatchers(HttpMethod.POST, "/users").hasAuthority("users.create")
+			.requestMatchers(HttpMethod.GET, "/users/**").hasAuthority("users.read")
+			.requestMatchers(HttpMethod.PUT, "/users/{username}").hasAuthority("users.update")
+			.requestMatchers(HttpMethod.DELETE, "/users/**").hasAuthority("users.delete")
+			// user setting
+			.requestMatchers("/usersettings/**").authenticated()
+			// pregnanttreatmenttypes
+			.requestMatchers(HttpMethod.POST, "/pregnanttreatmenttypes/**").hasAuthority("pregnanttreatmenttypes.create")
+			.requestMatchers(HttpMethod.GET, "/pregnanttreatmenttypes/**").hasAnyAuthority("pregnanttreatmenttypes.read")
+			.requestMatchers(HttpMethod.PUT, "/pregnanttreatmenttypes/**").hasAuthority("pregnanttreatmenttypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/pregnanttreatmenttypes/**").hasAuthority("pregnanttreatmenttypes.delete")
+			// pricelists
+			.requestMatchers(HttpMethod.POST, "/pricelists/**").hasAuthority("pricelists.create")
+			.requestMatchers(HttpMethod.GET, "/pricelists/**").hasAnyAuthority("pricelists.read")
+			.requestMatchers(HttpMethod.PUT, "/pricelists/**").hasAuthority("pricelists.update")
+			.requestMatchers(HttpMethod.DELETE, "/pricelists/**").hasAuthority("pricelists.delete")
+			// pricesothers
+			.requestMatchers(HttpMethod.POST, "/pricesothers/**").hasAuthority("pricesothers.create")
+			.requestMatchers(HttpMethod.GET, "/pricesothers/**").hasAnyAuthority("pricesothers.read")
+			.requestMatchers(HttpMethod.PUT, "/pricesothers/**").hasAuthority("pricesothers.update")
+			.requestMatchers(HttpMethod.DELETE, "/pricesothers/**").hasAuthority("pricesothers.delete")
+			// reports
+			.requestMatchers(HttpMethod.POST, "/reports/**").hasAuthority("reports.create")
+			.requestMatchers(HttpMethod.GET, "/reports/**").hasAnyAuthority("reports.read")
+			.requestMatchers(HttpMethod.PUT, "/reports/**").hasAuthority("reports.update")
+			.requestMatchers(HttpMethod.DELETE, "/reports/**").hasAuthority("reports.delete")
+			// sms
+			.requestMatchers(HttpMethod.POST, "/sms/**").hasAuthority("sms.create")
+			.requestMatchers(HttpMethod.GET, "/sms/**").hasAnyAuthority("sms.read")
+			.requestMatchers(HttpMethod.PUT, "/sms/**").hasAuthority("sms.update")
+			.requestMatchers(HttpMethod.DELETE, "/sms/**").hasAuthority("sms.delete")
+			// suppliers
+			.requestMatchers(HttpMethod.POST, "/suppliers/**").hasAuthority("suppliers.create")
+			.requestMatchers(HttpMethod.GET, "/suppliers/**").hasAnyAuthority("suppliers.read")
+			.requestMatchers(HttpMethod.PUT, "/suppliers/**").hasAuthority("suppliers.update")
+			.requestMatchers(HttpMethod.DELETE, "/suppliers/**").hasAuthority("suppliers.delete")
+			// therapy
+			.requestMatchers(HttpMethod.POST, "/therapies/**").hasAuthority("therapies.create")
+			.requestMatchers(HttpMethod.GET, "/therapies/**").hasAnyAuthority("therapies.read")
+			.requestMatchers(HttpMethod.PUT, "/therapies/**").hasAuthority("therapies.update")
+			.requestMatchers(HttpMethod.DELETE, "/therapies/**").hasAuthority("therapies.delete")
+			// vaccines
+			.requestMatchers(HttpMethod.POST, "/vaccines/**").hasAuthority("vaccines.create")
+			.requestMatchers(HttpMethod.GET, "/vaccines/**").hasAnyAuthority("vaccines.read")
+			.requestMatchers(HttpMethod.PUT, "/vaccines/**").hasAuthority("vaccines.update")
+			.requestMatchers(HttpMethod.DELETE, "/vaccines/**").hasAuthority("vaccines.delete")
+			// vaccineType
+			.requestMatchers(HttpMethod.POST, "/vaccinetypes/**").hasAuthority("vaccinetypes.create")
+			.requestMatchers(HttpMethod.GET, "/vaccinetypes/**").hasAnyAuthority("vaccinetypes.read")
+			.requestMatchers(HttpMethod.PUT, "/vaccinetypes/**").hasAuthority("vaccinetypes.update")
+			.requestMatchers(HttpMethod.DELETE, "/vaccinetypes/**").hasAuthority("vaccinetypes.delete")
+			// visit
+			.requestMatchers(HttpMethod.POST, "/visits/**").hasAuthority("visits.create")
+			.requestMatchers(HttpMethod.GET, "/visits/**").hasAnyAuthority("visits.read")
+			.requestMatchers(HttpMethod.PUT, "/visits/**").hasAuthority("visits.update")
+			.requestMatchers(HttpMethod.DELETE, "/visits/**").hasAuthority("visits.delete")
+			// wards
+			.requestMatchers(HttpMethod.POST, "/wards/**").hasAuthority("wards.create")
+			.requestMatchers(HttpMethod.GET, "/wards/**").hasAnyAuthority("wards.read")
+			.requestMatchers(HttpMethod.PUT, "/wards/**").hasAuthority("wards.update")
+			.requestMatchers(HttpMethod.DELETE, "/wards/**").hasAuthority("wards.delete")
 
-						// .requestMatchers("/auth-needed/**").authenticated()
-						// .requestMatchers("/noauth-public/**").permitAll()
-						// .requestMatchers("/admin/**").hasAuthority("admin")
+			// .requestMatchers("/auth-needed/**").authenticated()
+			// .requestMatchers("/noauth-public/**").permitAll()
+			// .requestMatchers("/admin/**").hasAuthority("admin")
 
-//			.and()
-//			.formLogin()
-//				 .loginPage("/auth/login")
-//				 .successHandler(successHandler())
-//				 .failureHandler(failureHandler())
-						.and().apply(securityConfigurerAdapter())
-						.and().httpBasic()
-						.and().logout().logoutUrl("/auth/logout").addLogoutHandler(customLogoutHandler).permitAll();
+			//			.and()
+			//			.formLogin()
+			//				 .loginPage("/auth/login")
+			//				 .successHandler(successHandler())
+			//				 .failureHandler(failureHandler())
+			.and().apply(securityConfigurerAdapter())
+			.and().httpBasic()
+			.and().logout().logoutUrl("/auth/logout").addLogoutHandler(customLogoutHandler).permitAll();
 		return http.build();
 	}
