+++ src/main/java/org/isf/security/CustomAuthenticationManager.java
@@ -22,35 +22,31 @@
 package org.isf.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Component
 public class CustomAuthenticationManager implements AuthenticationManager {
 
-    @Autowired
-    private UserDetailsServiceImpl customUserDetailsService;
+	@Autowired
+	private UserDetailsServiceImpl customUserDetailsService;
 
-    @Bean
-    protected PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
+	@Autowired
+	private PasswordEncoder passwordEncoder;
 
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        final UserDetails userDetail = customUserDetailsService.loadUserByUsername(authentication.getName());
-        if (!passwordEncoder().matches(authentication.getCredentials().toString(), userDetail.getPassword())) {
-            throw new BadCredentialsException("Wrong password");
-        }
-        return new UsernamePasswordAuthenticationToken(userDetail.getUsername(), userDetail.getPassword(), userDetail.getAuthorities());
-    }
+	@Override
+	public Authentication authenticate(Authentication authentication) throws AuthenticationException {
+		final UserDetails userDetail = customUserDetailsService.loadUserByUsername(authentication.getName());
+		if (!passwordEncoder.matches(authentication.getCredentials().toString(), userDetail.getPassword())) {
+			throw new BadCredentialsException("Wrong password");
+		}
+		return new UsernamePasswordAuthenticationToken(userDetail.getUsername(), userDetail.getPassword(), userDetail.getAuthorities());
+	}
 
 }
