+++ src/test/java/org/isf/usergroups/data/UserGroupHelper.java
@@ -19,14 +19,15 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.isf.menu.data;
+package org.isf.usergroups.data;
 
 import java.util.List;
 
 import org.isf.menu.TestUserGroup;
-import org.isf.menu.dto.UserGroupDTO;
 import org.isf.menu.model.UserGroup;
+import org.isf.usergroups.dto.UserGroupDTO;
 import org.isf.utils.exception.OHException;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
@@ -35,11 +36,11 @@ import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 
 /**
  * Helper class to generate DTOs and Entities for users endpoints test
- *
  * @author Silevester D.
  * @since 1.15
  */
 public class UserGroupHelper {
+
 	public static UserGroup generateUserGroup() throws OHException {
 		return new TestUserGroup().setup(false);
 	}
@@ -47,10 +48,10 @@ public class UserGroupHelper {
 	public static String asJsonString(UserGroupDTO userGroupDTO) {
 		try {
 			return new ObjectMapper()
-					.registerModule(new ParameterNamesModule())
-					.registerModule(new Jdk8Module())
-					.registerModule(new JavaTimeModule())
-					.writeValueAsString(userGroupDTO);
+							.registerModule(new ParameterNamesModule())
+							.registerModule(new Jdk8Module())
+							.registerModule(new JavaTimeModule())
+							.writeValueAsString(userGroupDTO);
 		} catch (JsonProcessingException e) {
 			e.printStackTrace();
 		}
@@ -61,10 +62,10 @@ public class UserGroupHelper {
 	public static String asJsonString(List<UserGroupDTO> userGroupDTOs) {
 		try {
 			return new ObjectMapper()
-					.registerModule(new ParameterNamesModule())
-					.registerModule(new Jdk8Module())
-					.registerModule(new JavaTimeModule())
-					.writeValueAsString(userGroupDTOs);
+							.registerModule(new ParameterNamesModule())
+							.registerModule(new Jdk8Module())
+							.registerModule(new JavaTimeModule())
+							.writeValueAsString(userGroupDTOs);
 		} catch (JsonProcessingException e) {
 			e.printStackTrace();
 		}
