+++ src/test/java/org/isf/users/data/UserHelper.java
@@ -19,15 +19,16 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.isf.menu.data;
+package org.isf.users.data;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import org.isf.menu.dto.UserDTO;
 import org.isf.menu.model.User;
 import org.isf.menu.model.UserGroup;
+import org.isf.usergroups.data.UserGroupHelper;
+import org.isf.users.dto.UserDTO;
 import org.isf.utils.exception.OHException;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -38,11 +39,11 @@ import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 
 /**
  * Helper class to generate DTOs and Entities for users endpoints test
- *
  * @author Silevester D.
  * @since 1.15
  */
 public class UserHelper {
+
 	public static User generateUser() throws OHException {
 		User user = new User();
 		user.setUserGroupName(UserGroupHelper.generateUserGroup());
@@ -68,10 +69,10 @@ public class UserHelper {
 	public static String asJsonString(UserDTO userDTO) {
 		try {
 			return new ObjectMapper()
-					.registerModule(new ParameterNamesModule())
-					.registerModule(new Jdk8Module())
-					.registerModule(new JavaTimeModule())
-					.writeValueAsString(userDTO);
+							.registerModule(new ParameterNamesModule())
+							.registerModule(new Jdk8Module())
+							.registerModule(new JavaTimeModule())
+							.writeValueAsString(userDTO);
 		} catch (JsonProcessingException e) {
 			e.printStackTrace();
 		}
@@ -82,10 +83,10 @@ public class UserHelper {
 	public static String asJsonString(List<UserDTO> users) {
 		try {
 			return new ObjectMapper()
-					.registerModule(new ParameterNamesModule())
-					.registerModule(new Jdk8Module())
-					.registerModule(new JavaTimeModule())
-					.writeValueAsString(users);
+							.registerModule(new ParameterNamesModule())
+							.registerModule(new Jdk8Module())
+							.registerModule(new JavaTimeModule())
+							.writeValueAsString(users);
 		} catch (JsonProcessingException e) {
 			e.printStackTrace();
 		}
