+++ src/test/resources/application.properties
@@ -1,6 +1,6 @@
 api.protocol=http
 api.host=localhost
-
+###
 spring.main.allow-bean-definition-overriding=true
 ### Security token secret (JWT)
 jwt.token.secret=+Z0E+6QXT9ri0L+PVkRTSF/tsZ/Z1yXJdrqz8FieiebiezOuIcuabJExyxCZi1bzSteaSmbCdzIaLDLPht92Bg==
